+++ src/simple_mc_main1.rs
@@ -8,7 +8,7 @@ fn simple_montecarlo1(
     spot: f64,
     vol: f64,
     r: f64,
-    number_of_paths: u32,
+    number_of_paths: u64,
 ) -> f64 {
     let variance = vol * vol * expiry;
     let root_variance = variance.sqrt();
